+++ types/kafkajs.d.ts
@@ -311,6 +311,10 @@ export interface OffsetsByTopicPartition {
   topics: TopicOffsets[]
 }
 
+export type FetchOffsetsPartition = PartitionOffset & { metadata: string | null, leaderEpoch: number | null, error?: LibrdKafkaError };  
+
+export type TopicInput = string[] | { topic: string; partitions: number[] }[]
+
 export type Consumer = Client & {
   subscribe(subscription: ConsumerSubscribeTopics | ConsumerSubscribeTopic): Promise<void>
   stop(): Promise<void>
@@ -369,4 +373,10 @@ export type Admin = {
     groups: string[],
     options?: { timeout?: number, includeAuthorizedOperations?: boolean }): Promise<GroupDescriptions>
   deleteGroups(groupIds: string[], options?: { timeout?: number }): Promise<DeleteGroupsResult[]>
+  fetchOffsets(options: { 
+    groupId: string,
+    topics?: TopicInput,
+    timeout?: number,
+    requireStableOffsets?: boolean }): 
+    Promise<Array<{topic: string; partitions:FetchOffsetsPartition[]}>>
 }