+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-# confluent-kafka-javascript v0.3.1
+# confluent-kafka-javascript v0.4.0
 
-v0.3.1 is a limited availability maintenance release. It is supported for all usage.
+v0.4.0 is a limited availability feature release. It is supported for all usage.
 
 ## Enhancements
 
 1. Fixes an issue where headers were not passed correctly to the `eachBatch` callback (#130).
+2. Add support for an Admin API to list a consumer group's offsets (#49).
 
 
 # confluent-kafka-javascript v0.3.0